+++ src/util/system.cpp
@@ -1235,7 +1235,7 @@ void RenameThreadPool(ctpl::thread_pool& tp, const char* baseName)
         // `doneCnt` should be at least `futures.size()` if tp size was increased (for whatever reason),
         // or at least `tp.size()` if tp size was decreased and queue was cleared
         // (which can happen on `stop()` if we were not fast enough to get all jobs to their threads).
-    } while (doneCnt < futures.size() && doneCnt < tp.size());
+    } while (doneCnt < static_cast<int64_t>(futures.size()) && doneCnt < tp.size());
 
     cond->notify_all();
